+++ laterpay/signing.py
@@ -1,13 +1,16 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, print_function
 
-import functools
 import hashlib
 import hmac
 import time
+import warnings
+
+import six
+
+from six.moves.urllib.parse import parse_qsl, quote, urlencode, urlparse
 
 from . import compat
-
 
 ALLOWED_METHODS = ('GET', 'POST', 'PUT', 'DELETE', 'OPTIONS', 'HEAD')
 
@@ -34,9 +37,9 @@ def create_HMAC(HMAC_secret, *parts):
     This function should probably not be part of the public API, and thus will
     be deprecated in a future release to be replaced with a internal function.
     """
-    authcode = hmac.new(compat.b(HMAC_secret), digestmod=hashlib.sha224)
+    authcode = hmac.new(six.b(HMAC_secret), digestmod=hashlib.sha224)
     for part in parts:
-        authcode.update(compat.b(part))
+        authcode.update(six.b(part))
     return authcode.hexdigest()
 
 
@@ -47,27 +50,19 @@ def sort_params(param_dict):
     This function should probably not be part of the public API, and thus will
     be deprecated in a future release to be replaced with a internal function.
     """
-    def cmp_params(param1, param2):
-        result = compat.cmp(param1[0], param2[0])
-        if result == 0:
-            result = compat.cmp(param1[1], param2[1])
-        return result
-
     param_list = []
     for name, value_list in param_dict.items():
         if isinstance(value_list, (list, tuple)):
             for value in value_list:
-                if not isinstance(value, compat.string_types):
+                if not isinstance(value, six.string_types):
                     value = str(value)
                 param_list.append((name, value))
         else:
-            if not isinstance(value_list, compat.string_types):
+            if not isinstance(value_list, six.string_types):
                 value_list = str(value_list)
             param_list.append((name, value_list))
 
-    if compat.py3k:
-        return sorted(param_list, key=functools.cmp_to_key(cmp_params))
-    return sorted(param_list, cmp_params)
+    return sorted(param_list)
 
 
 def normalise_param_structure(params):
@@ -117,11 +112,11 @@ def create_base_message(params, url, method='POST'):
     """
     msg = '{method}&{url}&{params}'
 
-    method = _encode_if_unicode(method).upper()
+    method = compat.encode_if_unicode(method).upper()
 
     data = {}
 
-    url = compat.quote(_encode_if_unicode(url), safe='')
+    url = quote(compat.encode_if_unicode(url), safe='')
 
     if method not in ALLOWED_METHODS:
         raise ValueError('method should be one of: {}'.format(ALLOWED_METHODS))
@@ -129,49 +124,37 @@ def create_base_message(params, url, method='POST'):
     params = normalise_param_structure(params)
 
     for key, values in params.items():
-        key = compat.quote(_encode_if_unicode(key), safe='')
-
-        if not isinstance(values, (list, tuple)):
-            values = [values]
+        key = quote(compat.encode_if_unicode(key), safe='')
 
         values_str = []
 
-        # If any non compat.string_types objects, ``str()`` them.
+        # If any non six.string_types objects, ``str()`` them.
         for value in values:
-            if not isinstance(value, compat.string_types):
+            if not isinstance(value, six.string_types):
                 value = str(value)
             values_str.append(value)
 
-        data[key] = [compat.quote(_encode_if_unicode(value_str), safe='') for value_str in values_str]
+        data[key] = [quote(compat.encode_if_unicode(value_str), safe='') for value_str in values_str]
 
     sorted_params = sort_params(data)
 
     param_str = '&'.join('{}={}'.format(k, v) for k, v in sorted_params)
-    param_str = compat.quote(param_str, safe='')
+    param_str = quote(param_str, safe='')
 
     return msg.format(method=method, url=url, params=param_str)
 
 
 def sign(secret, params, url, method='POST'):
     """
-    Create signature for given `params`, `url` and HTTP method.
+    Create signature for given `params`, `url` and HTTP `method`.
 
-    How params are canonicalized:
-    - `compat.quote` every key and value that will be signed
-    - sort the param list
-    - `'&'join` the params
-
-    :param secret: Secret used to create signature
-    :type secret: str/unicode
-    :param params: Mapping of all params that should be signed.
-        This includes url params as well as requests body params,
-        regardless if body is 'application/x-www-form-urlencoded' or
-        'application/json'.
-    :type params: dict of unicode/str keys and values. Values can
-        be a list/tuple of unicode/str.
-    :param url: full url of the target endpoint, no url params
-        (eg. https://host/api)
-    :type url: str/unicode
+    :param secret: secret string used to create the signature
+    :param params: params dict (values can be strings or lists of strings)
+    :param url: base url for which the params were signed.
+                Example: https://example.net/here
+                (no query params or fragments)
+    :param method: HTTP method used to transport the signed data
+                   ('POST' is default)
     """
     if 'hmac' in params:
         params.pop('hmac')
@@ -179,9 +162,9 @@ def sign(secret, params, url, method='POST'):
     if 'gettoken' in params:
         params.pop('gettoken')
 
-    secret = _encode_if_unicode(secret)
+    secret = compat.encode_if_unicode(secret)
 
-    url_parsed = compat.urlparse(url)
+    url_parsed = urlparse(url)
     base_url = url_parsed.scheme + "://" + url_parsed.netloc + url_parsed.path
 
     msg = create_base_message(params, base_url, method=method)
@@ -193,10 +176,15 @@ def sign(secret, params, url, method='POST'):
 
 def verify(signature, secret, params, url, method):
     """
-    Verify the signature on a signed URL.
+    Verify the signature of a given `params` dict.
 
-    Redirects back to a client's server from LaterPay will be signed to ensure
-    authenticity. Use `verify` to confirm this.
+    :param signature: signature string to be verified
+    :param secret: secret string used to create the signature
+    :param params: params dict (values can be strings or lists of strings)
+    :param url: base url for which the params were signed.
+                Example: https://example.net/here
+                (no query params or fragments)
+    :param method: HTTP method used to transport the signed data
     """
     if isinstance(signature, (list, tuple)):
         signature = signature[0]
@@ -208,6 +196,8 @@ def verify(signature, secret, params, url, method):
 
 def sign_and_encode(secret, params, url, method="GET"):
     """
+    Deprecated. Consider using ``laterpay.utils.signed_query()`` instead.
+
     Sign and encode a URL ``url`` with a ``secret`` key called via an HTTP ``method``.
 
     It adds the signature to the URL
@@ -220,6 +210,12 @@ def sign_and_encode(secret, params, url, method="GET"):
     :param method: an uppercase string representation of the HTTP method being used for the call (e.g. "GET", "POST")
     :return: a signed and correctly encoded URL
     """
+    warnings.warn(
+        "sign_and_encode is deprecated. It will be removed in a future release. "
+        "Consider using ``laterpay.utils.signed_query()`` instead.",
+        DeprecationWarning,
+    )
+
     if 'ts' not in params:
         params['ts'] = str(int(time.time()))
 
@@ -228,11 +224,11 @@ def sign_and_encode(secret, params, url, method="GET"):
 
     sorted_data = []
     for k, v in sort_params(params):
-        k = _encode_if_unicode(k)
-        value = _encode_if_unicode(v)
+        k = compat.encode_if_unicode(k)
+        value = compat.encode_if_unicode(v)
         sorted_data.append((k, value))
 
-    encoded = compat.urlencode(sorted_data)
+    encoded = urlencode(sorted_data)
     hmac = sign(secret, params, url=url, method=method)
 
     return "%s&hmac=%s" % (encoded, hmac)
@@ -240,6 +236,8 @@ def sign_and_encode(secret, params, url, method="GET"):
 
 def sign_get_url(secret, url, signature_paramname="hmac"):
     """
+    Deprecated.
+
     Sign a URL to be GET-ed.
 
     This function takes a URL, parses it, sorts the URL parameters in
@@ -261,12 +259,19 @@ def sign_get_url(secret, url, signature_paramname="hmac"):
     :type signature_paramname: str
     :returns: ``str`` -- the URL, including the signature as an URL parameter
     """
-    parsed = compat.urlparse(url)
+    warnings.warn(
+        "sign_get_url is deprecated. It will be removed in a future release. "
+        "It wasn't intended for public use. It's recommended to use the core "
+        "signing API which is sign() and verify().",
+        DeprecationWarning,
+    )
+
+    parsed = urlparse(url)
 
     if parsed.query != "":
-        # use compat.parse_qsl, because .parse_qs seems to create problems
-        # with compat.urlencode()
-        qs = compat.parse_qsl(parsed.query, keep_blank_values=True)
+        # use parse_qsl, because .parse_qs seems to create problems
+        # with urlencode()
+        qs = parse_qsl(parsed.query, keep_blank_values=True)
 
         # create string to sign
 
@@ -278,21 +283,6 @@ def sign_get_url(secret, url, signature_paramname="hmac"):
 
         qs.append((signature_paramname, hmac))
         return parsed.scheme + "://" + parsed.netloc + parsed.path + \
-            parsed.params + "?" + compat.urlencode(qs) + parsed.fragment
+            parsed.params + "?" + urlencode(qs) + parsed.fragment
 
     return None
-
-
-def _encode_if_unicode(value, encoding='utf-8'):
-    """
-    Encode and return a ``value`` using specified ``encoding``.
-
-    Encoding is done only if ``value`` is a ``unicode`` instance
-    (utf-8 encoding is used as default).
-
-    This utility is needed because some web frameworks can provide
-    request arguments as ``str`` instances.
-    """
-    if not compat.py3k and isinstance(value, unicode):
-        value = value.encode(encoding)
-    return value