+++ CHANGELOG.md
@@ -1,8 +1,32 @@
 # Changelog
 
+
 ## Unreleased
 
-TBD
+
+## 4.2.0
+
+* `laterpay.signing.sign_and_encode()` is deprecated and will be removed in a
+  future release. Consider using `laterpay.utils.signed_query()` instead.
+* `signing.sign_get_url()` is deprecated and will be removed in a future
+  release.
+* `LaterPayClient.get_access()` is deprecated and will be removed in a future
+  release. Consider using `LaterPayClient.get_access_data()` instead.
+* `LaterPayClient.get_identify_url()` is deprecated and will be removed in a future
+  release.
+* New utility functions: `laterpay.utils.signed_query()` and
+  `laterpay.utils.signed_url()`
+* New `LaterPayClient` methods:
+
+    * `get_request_headers()`
+    * `get_access_url()`
+    * `get_access_params()`
+    * `get_access_data()`
+
+* Improved and newly added tests.
+* Improved `laterpay.signing` docs.
+* Replaced most of `compat` with proper `six` usage.
+
 
 ## 4.1.0